Wrapped the casting and testing of the long handle in preprocessor directives. This is needed when the SIZE_OF_VOID_P is not the same (usually greater than) the SIZE_OF_UNSIGNED_INT.
The typedef for pointertype in xcircuit.h using a
similar mechanism.

When SIZE_OF_VOID_P is the same as SIZE_OF_UNSIGNED_LONG, then
there is no need to cast the handle down and up, as pointertype
is the same as the long handle.

If this protection is not in place, data loss and an Application
Error can occur when the handle value is larger than an int.